Create `TrackedAsyncData` type and `load` helper
`TrackedAsyncData` is an autotracked/reactive handler for promises. It
provides public access to the state of the promise:

- a `state` property of `"PENDING"`, `"RESOLVED"`, or `"REJECTED"`
- boolean accessors (useful especially in templates) for those states:
  `isPending`, `isResolved, and `isRejected`
- for a resolved promise, the resolved `value`
- for a rejected promise, the rejection `error`

It is backwards compatible with Ember Classic usage via classic computed
properties, but requires Ember Octane and will warn about accesses that
will be deprecated in a future major version (probably around Ember 4.0)
which will drops support for Ember Classic.