Improve type robustness for `TrackedAsyncData<T>`
Export a union type as the actual public API of `TrackedAsyncData<T>`,
along with a `const`-bound export of the class (with a safe cast for the
constructor). The union is implemented as a union of interfaces which
extend the base class so that (a) they are guaranteed to be structurally
compatible and (b) docstrings continue to "just work".

The upshot of this change: All existing invocations continue to work
exactly as they have up to this point, but the type now correctly
narrows when checking any of the boolean properties exposed to indicate
the state of the type. For example, given a `TrackedAsyncData<string>`,
the `.value` and `.error` properties will have the type `null` when
`.isPending` is `true`; `.value` will have the type `string` when
`.isResolved` is `true` but `.error` will still be `null`, etc.