Supply correct types for `state`
As with the `.isPending`, `.isResolved`, and `.isRejected` properties,
the value of the `.state` property is known statically, so the exposed
types can treat it as a static string.

Additionally, since end users may depend on it in Ember Classic code
with dependent keys, like `@computed('someAsyncData.state')`, decorate
it with `@dependentKeyCompat`.