Fix `paths` location in type tests tsconfig.json
`paths` should be in the `compilerOptions` hash, and should map from the
root of the repo. It should include *only* the addon (not tests) and the
path for the destroyable polyfill. In the future, we should revert to
using `ember ts:precompile` and look things up from the root again, but
this works to unblock us for now.